First commited as before_9451_rev3

+++ src/main/java/net/spy/memcached/protocol/binary/OperationImpl.java
@@ -73,7 +73,7 @@ abstract class OperationImpl extends BaseOperationImpl implements Operation {
   private static final AtomicInteger SEQ_NUMBER = new AtomicInteger(0);
 
   // request header fields
-  private final byte cmd;
+  private final int cmd;
   protected short vbucket = 0;
   protected final int opaque;
 
@@ -96,7 +96,7 @@ abstract class OperationImpl extends BaseOperationImpl implements Operation {
    * @param o the opaque value.
    * @param cb
    */
-  protected OperationImpl(byte c, int o, OperationCallback cb) {
+  protected OperationImpl(int c, int o, OperationCallback cb) {
     super();
     cmd = c;
     opaque = o;
@@ -307,7 +307,7 @@ abstract class OperationImpl extends BaseOperationImpl implements Operation {
     ByteBuffer bb = ByteBuffer.allocate(bufSize + extraLen);
     assert bb.order() == ByteOrder.BIG_ENDIAN;
     bb.put(REQ_MAGIC);
-    bb.put(cmd);
+    bb.put((byte) cmd);
     bb.putShort((short) keyBytes.length);
     bb.put((byte) extraLen);
     bb.put((byte) 0); // data type
